Clicking on a dot no longer creates a new dot
If you have added a couple of dots and then click on an existing
dot chances are you want a line to that particular dot and not a
new dot almost on top of it. Fixed so that no new dot is added if
you click on an existing one.